requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.6.0
ete3==3.1.2
idna==3.1
multidict==5.1.0
numpy==1.20.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
